shuffle some bits around
