+++ src/mame/tatsumi/tatsumi.h
@@ -77,75 +77,4 @@ protected:
 };
 
 
-
-class cyclwarr_state : public tatsumi_state
-{
-public:
-	cyclwarr_state(const machine_config &mconfig, device_type type, const char *tag)
-		: tatsumi_state(mconfig, type, tag)
-		, m_soundlatch(*this, "soundlatch")
-		, m_master_ram(*this, "master_ram")
-		, m_slave_ram(*this, "slave_ram")
-		, m_cyclwarr_videoram(*this, "cw_videoram%u", 0U)
-		, m_cyclwarr_tileclut(*this, "cw_tileclut")
-	{
-	}
-
-	void cyclwarr(machine_config &config);
-	void bigfight(machine_config &config);
-
-protected:
-	virtual void machine_reset() override ATTR_COLD;
-
-private:
-	uint16_t cyclwarr_sprite_r(offs_t offset);
-	void cyclwarr_sprite_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
-	void video_config_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
-	void bigfight_a40000_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
-	void mixing_control_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
-	void cyclwarr_control_w(uint8_t data);
-	void cyclwarr_sound_w(uint8_t data);
-	void output_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
-	uint8_t oki_status_xor_r();
-
-	template<int Bank> uint16_t cyclwarr_videoram_r(offs_t offset);
-	template<int Bank> void cyclwarr_videoram_w(offs_t offset, uint16_t data, uint16_t mem_mask = ~0);
-
-	template<int Bank> TILE_GET_INFO_MEMBER(get_tile_info_bigfight);
-	template<int Bank> TILE_GET_INFO_MEMBER(get_tile_info_cyclwarr_road);
-	DECLARE_VIDEO_START(cyclwarr);
-	DECLARE_VIDEO_START(bigfight);
-	uint32_t screen_update_cyclwarr(screen_device &screen, bitmap_rgb32 &bitmap, const rectangle &cliprect);
-
-	void common_map(address_map &map) ATTR_COLD;
-	void master_map(address_map &map) ATTR_COLD;
-	void slave_map(address_map &map) ATTR_COLD;
-	void sound_map(address_map &map) ATTR_COLD;
-
-	required_device<generic_latch_8_device> m_soundlatch;
-
-	required_shared_ptr<uint16_t> m_master_ram;
-	required_shared_ptr<uint16_t> m_slave_ram;
-	required_shared_ptr_array<uint16_t, 2> m_cyclwarr_videoram;
-	required_region_ptr<uint8_t> m_cyclwarr_tileclut;
-
-	std::vector<uint8_t> m_mask;
-	tilemap_t *m_layer[4]{};
-
-	uint16_t m_video_config[4]{};
-	uint16_t m_mixing_control = 0;
-	uint16_t m_bigfight_a40000[2]{};
-	uint16_t m_bigfight_bank = 0;
-	uint16_t m_bigfight_last_bank = 0;
-	uint16_t m_road_color_bank = 0, m_prev_road_bank = 0;
-	uint16_t m_layer_page_size[4]{};
-	bool m_layer1_can_be_road = false;
-	std::unique_ptr<uint8_t[]> m_decoded_gfx;
-
-	void tile_expand();
-	void draw_bg(screen_device &screen, bitmap_rgb32 &bitmap, const rectangle &cliprect, tilemap_t *src, const uint16_t* scrollx, const uint16_t* scrolly, const uint16_t layer_page_size, bool is_road, int hi_priority);
-	void draw_bg_layers(screen_device &screen, bitmap_rgb32 &bitmap, const rectangle &cliprect, int hi_priority);
-	void apply_highlight_bitmap(bitmap_rgb32 &bitmap, const rectangle &cliprect, bitmap_ind8 &highlight_bitmap);
-};
-
 #endif // MAME_TATSUMI_TATSUMI_H
