--- src/mame/tatsumi/tatsumi_m.cpp
@@ -1,123 +1,0 @@
-// license:BSD-3-Clause
-// copyright-holders:Bryan McPhail, Angelo Salese
-#include "emu.h"
-#include "tatsumi.h"
-#include "sound/okim6295.h"
-
-
-/******************************************************************************/
-
-void tatsumi_state::tatsumi_reset()
-{
-	m_last_control = 0;
-	m_control_word = 0;
-
-	save_item(NAME(m_last_control));
-	save_item(NAME(m_control_word));
-}
-
-/******************************************************************************/
-
-void cyclwarr_state::cyclwarr_control_w(uint8_t data)
-{
-	m_control_word = data;
-
-//  if ((m_control_word&0xfe) != (m_last_control&0xfe))
-//      logerror("%s  control_w %04x\n", m_maincpu->pc(), data);
-
-/*
-
-0x1 - watchdog
-0x4 - cpu bus lock
-
-
-
-*/
-
-	if ((m_control_word & 4) == 4 && (m_last_control & 4) == 0)
-	{
-//      logerror("68k 2 halt\n");
-		m_subcpu->set_input_line(INPUT_LINE_HALT, ASSERT_LINE);
-	}
-
-	if ((m_control_word & 4) == 0 && (m_last_control & 4) == 4)
-	{
-//      logerror("68k 2 irq go\n");
-		m_subcpu->set_input_line(INPUT_LINE_HALT, CLEAR_LINE);
-	}
-
-	m_last_control = m_control_word;
-}
-
-/******************************************************************************/
-
-uint16_t tatsumi_state::tatsumi_v30_68000_r(offs_t offset)
-{
-	const uint16_t* rom=(uint16_t*)m_subregion->base();
-
-//logerror("%s:68000_r(%04X),cw=%04X\n", m_maincpu->pc(), offset*2, m_control_word);
-	/* Read from 68k RAM */
-	if ((m_control_word&0x1f)==0x18)
-	{
-#if 0
-		// hack to make roundup 5 boot
-		// doesn't seem necessary anymore, left for reference
-		if (m_maincpu->pc()==0xec575)
-		{
-			uint8_t *dst = m_mainregion->base();
-			dst[BYTE_XOR_LE(0xec57a)]=0x46;
-			dst[BYTE_XOR_LE(0xec57b)]=0x46;
-
-			dst[BYTE_XOR_LE(0xfc520)]=0x46; //code that stops cpu after coin counter goes mad..
-			dst[BYTE_XOR_LE(0xfc521)]=0x46;
-			dst[BYTE_XOR_LE(0xfc522)]=0x46;
-			dst[BYTE_XOR_LE(0xfc523)]=0x46;
-			dst[BYTE_XOR_LE(0xfc524)]=0x46;
-			dst[BYTE_XOR_LE(0xfc525)]=0x46;
-		}
-#endif
-
-		return m_sharedram[offset & 0x1fff];
-	}
-
-	/* Read from 68k ROM */
-	offset+=(m_control_word&0x7)*0x8000;
-
-	return rom[offset];
-}
-
-void tatsumi_state::tatsumi_v30_68000_w(offs_t offset, uint16_t data, uint16_t mem_mask)
-{
-	if ((m_control_word&0x1f)!=0x18)
-		logerror("68k write in bank %05x\n",m_control_word);
-
-	COMBINE_DATA(&m_sharedram[offset]);
-}
-
-/***********************************************************************************/
-
-uint8_t cyclwarr_state::oki_status_xor_r()
-{
-	int r = m_oki->read();
-
-	// Cycle Warriors and Big Fight access this with reversed activeness.
-	// this is particularly noticeable with the "We got em" sample played in CW at stage clear:
-	// gets cut too early with the old hack below.
-	// fwiw returning normal oki status doesn't work at all, both games don't make any sound.
-	// TODO: verify with HW
-	return (r ^ 0xff);
-#if 0
-	// old hack left for reference
-
-	if (m_audiocpu->pc()==0x2b70 || m_audiocpu->pc()==0x2bb5
-		|| m_audiocpu->pc()==0x2acc
-		|| m_audiocpu->pc()==0x1c79 // BigFight
-		|| m_audiocpu->pc()==0x1cbe // BigFight
-		|| m_audiocpu->pc()==0xf9881)
-		return 0xf;
-	if (m_audiocpu->pc()==0x2ba3 || m_audiocpu->pc()==0x2a9b || m_audiocpu->pc()==0x2adc
-		|| m_audiocpu->pc()==0x1cac) // BigFight
-		return 0;
-	return r;
-#endif
-}
